+++ modules/apps/foundation/websocket/websocket-whiteboard/src/main/java/com/liferay/websocket/whiteboard/internal/WebSocketEndpointTracker.java
@@ -14,14 +14,18 @@
 
 package com.liferay.websocket.whiteboard.internal;
 
+import java.util.List;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
 import javax.servlet.ServletContext;
 
+import javax.websocket.Decoder;
 import javax.websocket.DeploymentException;
+import javax.websocket.Encoder;
 import javax.websocket.Endpoint;
 import javax.websocket.server.ServerContainer;
+import javax.websocket.server.ServerEndpointConfig;
 
 import org.osgi.framework.BundleContext;
 import org.osgi.framework.ServiceObjects;
@@ -53,6 +57,15 @@ public class WebSocketEndpointTracker
 			return null;
 		}
 
+		List<Class<? extends Decoder>> decoders =
+			(List<Class<? extends Decoder>>)serviceReference.getProperty(
+				"org.osgi.http.websocket.endpoint.decoders");
+		List<Class<? extends Encoder>> encoders =
+			(List<Class<? extends Encoder>>)serviceReference.getProperty(
+				"org.osgi.http.websocket.endpoint.encoders");
+		List<String> subprotocol = (List<String>)serviceReference.getProperty(
+			"org.osgi.http.websocket.endpoint.subprotocol");
+
 		final ServiceObjects<Endpoint> serviceObjects =
 			_bundleContext.getServiceObjects(serviceReference);
 
@@ -63,9 +76,39 @@ public class WebSocketEndpointTracker
 
 		if (serverEndpointConfigWrapper == null) {
 			serverEndpointConfigWrapper = new ServerEndpointConfigWrapper(
-				path, _logService);
+				path, decoders, encoders, subprotocol, _logService);
 
 			isNew = true;
+		}
+		else {
+			Class<?> endpointClass =
+				serverEndpointConfigWrapper.getEndpointClass();
+
+			ServerEndpointConfig.Configurator configurator =
+				serverEndpointConfigWrapper.getConfigurator();
+
+			try {
+				Object endpointInstance = configurator.getEndpointInstance(
+					endpointClass);
+
+				Class<?> endpointInstanceClass = endpointInstance.getClass();
+
+				if (endpointInstanceClass.equals(
+						ServerEndpointConfigWrapper.NullEndpoint.class)) {
+
+					serverEndpointConfigWrapper.override(
+						decoders, encoders, subprotocol);
+				}
+			}
+			catch (InstantiationException ie) {
+				Endpoint endpoint = serviceObjects.getService();
+
+				_logService.log(
+					LogService.LOG_ERROR,
+					"Unable to register WebSocket endpoint " +
+						endpoint.getClass() + " for path " + path,
+					ie);
+			}
 		}
 
 		serverEndpointConfigWrapper.setConfigurator(