+++ modules/sdk/gradle-plugins-defaults/src/main/java/com/liferay/gradle/plugins/defaults/internal/NodeDefaultsPlugin.java
@@ -18,6 +18,9 @@ import com.liferay.gradle.plugins.BaseDefaultsPlugin;
 import com.liferay.gradle.plugins.defaults.internal.util.GradleUtil;
 import com.liferay.gradle.plugins.node.NodeExtension;
 import com.liferay.gradle.plugins.node.NodePlugin;
+import com.liferay.gradle.plugins.node.tasks.NpmShrinkwrapTask;
+
+import java.util.Collections;
 
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
@@ -31,10 +34,8 @@ public class NodeDefaultsPlugin extends BaseDefaultsPlugin<NodePlugin> {
 
 	@Override
 	protected void configureDefaults(Project project, NodePlugin nodePlugin) {
-		NodeExtension nodeExtension = GradleUtil.getExtension(
-			project, NodeExtension.class);
-
-		nodeExtension.setNodeVersion(_NODE_VERSION);
+		_configureNode(project);
+		_configureTaskNpmShrinkwrap(project);
 	}
 
 	@Override
@@ -45,6 +46,26 @@ public class NodeDefaultsPlugin extends BaseDefaultsPlugin<NodePlugin> {
 	private NodeDefaultsPlugin() {
 	}
 
+	private void _configureNode(Project project) {
+		NodeExtension nodeExtension = GradleUtil.getExtension(
+			project, NodeExtension.class);
+
+		nodeExtension.setNodeVersion(_NODE_VERSION);
+	}
+
+	private void _configureTaskNpmShrinkwrap(Project project) {
+		NpmShrinkwrapTask npmShrinkwrapTask =
+			(NpmShrinkwrapTask)GradleUtil.getTask(
+				project, NodePlugin.NPM_SHRINKWRAP_TASK_NAME);
+
+		npmShrinkwrapTask.excludeDependencies(
+			_NPM_SHRINKWRAP_EXCLUDED_DEPENDENCIES);
+	}
+
 	private static final String _NODE_VERSION = "6.6.0";
+
+	private static final Iterable<String>
+		_NPM_SHRINKWRAP_EXCLUDED_DEPENDENCIES = Collections.singleton(
+			"fsevents");
 
 }