+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/internal/LangBuilderDefaultsPlugin.java
@@ -20,6 +20,7 @@ import com.liferay.gradle.plugins.lang.builder.BuildLangTask;
 import com.liferay.gradle.plugins.lang.builder.LangBuilderPlugin;
 
 import org.gradle.api.Action;
+import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.tasks.TaskContainer;
 
@@ -29,42 +30,16 @@ import org.gradle.api.tasks.TaskContainer;
 public class LangBuilderDefaultsPlugin
 	extends BasePortalToolDefaultsPlugin<LangBuilderPlugin> {
 
+	public static final Plugin<Project> INSTANCE =
+		new LangBuilderDefaultsPlugin();
+
 	@Override
 	protected void configureDefaults(
 		Project project, LangBuilderPlugin langBuilderPlugin) {
 
 		super.configureDefaults(project, langBuilderPlugin);
 
-		configureTasksBuildLang(project);
-	}
-
-	protected void configureTaskBuildLang(BuildLangTask buildLangTask) {
-		String translateClientId = GradleUtil.getProperty(
-			buildLangTask.getProject(), "microsoft.translator.client.id",
-			(String)null);
-
-		buildLangTask.setTranslateClientId(translateClientId);
-
-		String translateClientSecret = GradleUtil.getProperty(
-			buildLangTask.getProject(), "microsoft.translator.client.secret",
-			(String)null);
-
-		buildLangTask.setTranslateClientSecret(translateClientSecret);
-	}
-
-	protected void configureTasksBuildLang(Project project) {
-		TaskContainer taskContainer = project.getTasks();
-
-		taskContainer.withType(
-			BuildLangTask.class,
-			new Action<BuildLangTask>() {
-
-				@Override
-				public void execute(BuildLangTask buildLangTask) {
-					configureTaskBuildLang(buildLangTask);
-				}
-
-			});
+		_configureTasksBuildLang(project);
 	}
 
 	@Override
@@ -80,6 +55,38 @@ public class LangBuilderDefaultsPlugin
 	@Override
 	protected String getPortalToolName() {
 		return _PORTAL_TOOL_NAME;
+	}
+
+	private LangBuilderDefaultsPlugin() {
+	}
+
+	private void _configureTaskBuildLang(BuildLangTask buildLangTask) {
+		String translateClientId = GradleUtil.getProperty(
+			buildLangTask.getProject(), "microsoft.translator.client.id",
+			(String)null);
+
+		buildLangTask.setTranslateClientId(translateClientId);
+
+		String translateClientSecret = GradleUtil.getProperty(
+			buildLangTask.getProject(), "microsoft.translator.client.secret",
+			(String)null);
+
+		buildLangTask.setTranslateClientSecret(translateClientSecret);
+	}
+
+	private void _configureTasksBuildLang(Project project) {
+		TaskContainer taskContainer = project.getTasks();
+
+		taskContainer.withType(
+			BuildLangTask.class,
+			new Action<BuildLangTask>() {
+
+				@Override
+				public void execute(BuildLangTask buildLangTask) {
+					_configureTaskBuildLang(buildLangTask);
+				}
+
+			});
 	}
 
 	private static final String _PORTAL_TOOL_NAME = "com.liferay.lang.builder";