+++ modules/util/portal-tools-service-builder/src/main/java/com/liferay/portal/tools/service/builder/ServiceBuilder.java
@@ -233,85 +233,174 @@ public class ServiceBuilder {
 				modifiedFileNames);
 		}
 		catch (Exception e) {
-			String message =
-				"Please set these arguments. Sample values are:\n" +
-				"\n" +
-				"\tservice.api.dir=${basedir}/../portal-kernel/src\n" +
-				"\tservice.auto.import.default.references=true\n" +
-				"\tservice.auto.namespace.tables=false\n" +
-				"\tservice.bean.locator.util=com.liferay.portal.kernel.bean.PortalBeanLocatorUtil\n" +
-				"\tservice.build.number=1\n" +
-				"\tservice.build.number.increment=true\n" +
-				"\tservice.hbm.file=${basedir}/src/META-INF/portal-hbm.xml\n" +
-				"\tservice.impl.dir=${basedir}/src\n" +
-				"\tservice.input.file=${service.file}\n" +
-				"\tservice.model.hints.configs=" + StringUtil.merge(ServiceBuilderArgs.MODEL_HINTS_CONFIGS) + "\n" +
-				"\tservice.model.hints.file=${basedir}/src/META-INF/portal-model-hints.xml\n" +
-				"\tservice.osgi.module=false\n" +
-				"\tservice.plugin.name=\n" +
-				"\tservice.props.util=com.liferay.portal.util.PropsUtil\n" +
-				"\tservice.read.only.prefixes=" + StringUtil.merge(ServiceBuilderArgs.READ_ONLY_PREFIXES) + "\n" +
-				"\tservice.resource.actions.configs=" + StringUtil.merge(ServiceBuilderArgs.RESOURCE_ACTION_CONFIGS) + "\n" +
-				"\tservice.resources.dir=${basedir}/src\n" +
-				"\tservice.spring.file=${basedir}/src/META-INF/portal-spring.xml\n" +
-				"\tservice.spring.namespaces=beans\n" +
-				"\tservice.sql.dir=${basedir}/../sql\n" +
-				"\tservice.sql.file=portal-tables.sql\n" +
-				"\tservice.sql.indexes.file=indexes.sql\n" +
-				"\tservice.sql.sequences.file=sequences.sql\n" +
-				"\tservice.target.entity.name=${service.target.entity.name}\n" +
-				"\tservice.test.dir=${basedir}/test/integration\n" +
-				"\n" +
-				"You can also customize the generated code by overriding the default templates with these optional system properties:\n" +
-				"\n" +
-				"\t-Dservice.tpl.bad_alias_names=" + _TPL_ROOT + "bad_alias_names.txt\n" +
-				"\t-Dservice.tpl.bad_column_names=" + _TPL_ROOT + "bad_column_names.txt\n" +
-				"\t-Dservice.tpl.bad_json_types=" + _TPL_ROOT + "bad_json_types.txt\n" +
-				"\t-Dservice.tpl.bad_table_names=" + _TPL_ROOT + "bad_table_names.txt\n" +
-				"\t-Dservice.tpl.base_mode_impl=" + _TPL_ROOT + "base_mode_impl.ftl\n" +
-				"\t-Dservice.tpl.blob_model=" + _TPL_ROOT + "blob_model.ftl\n" +
-				"\t-Dservice.tpl.copyright.txt=copyright.txt\n" +
-				"\t-Dservice.tpl.ejb_pk=" + _TPL_ROOT + "ejb_pk.ftl\n" +
-				"\t-Dservice.tpl.exception=" + _TPL_ROOT + "exception.ftl\n" +
-				"\t-Dservice.tpl.extended_model=" + _TPL_ROOT + "extended_model.ftl\n" +
-				"\t-Dservice.tpl.extended_model_base_impl=" + _TPL_ROOT + "extended_model_base_impl.ftl\n" +
-				"\t-Dservice.tpl.extended_model_impl=" + _TPL_ROOT + "extended_model_impl.ftl\n" +
-				"\t-Dservice.tpl.finder=" + _TPL_ROOT + "finder.ftl\n" +
-				"\t-Dservice.tpl.finder_base_impl=" + _TPL_ROOT + "finder_base_impl.ftl\n" +
-				"\t-Dservice.tpl.finder_util=" + _TPL_ROOT + "finder_util.ftl\n" +
-				"\t-Dservice.tpl.hbm_xml=" + _TPL_ROOT + "hbm_xml.ftl\n" +
-				"\t-Dservice.tpl.json_js=" + _TPL_ROOT + "json_js.ftl\n" +
-				"\t-Dservice.tpl.json_js_method=" + _TPL_ROOT + "json_js_method.ftl\n" +
-				"\t-Dservice.tpl.model=" + _TPL_ROOT + "model.ftl\n" +
-				"\t-Dservice.tpl.model_cache=" + _TPL_ROOT + "model_cache.ftl\n" +
-				"\t-Dservice.tpl.model_hints_xml=" + _TPL_ROOT + "model_hints_xml.ftl\n" +
-				"\t-Dservice.tpl.model_impl=" + _TPL_ROOT + "model_impl.ftl\n" +
-				"\t-Dservice.tpl.model_soap=" + _TPL_ROOT + "model_soap.ftl\n" +
-				"\t-Dservice.tpl.model_wrapper=" + _TPL_ROOT + "model_wrapper.ftl\n" +
-				"\t-Dservice.tpl.persistence=" + _TPL_ROOT + "persistence.ftl\n" +
-				"\t-Dservice.tpl.persistence_impl=" + _TPL_ROOT + "persistence_impl.ftl\n" +
-				"\t-Dservice.tpl.persistence_util=" + _TPL_ROOT + "persistence_util.ftl\n" +
-				"\t-Dservice.tpl.props=" + _TPL_ROOT + "props.ftl\n" +
-				"\t-Dservice.tpl.service=" + _TPL_ROOT + "service.ftl\n" +
-				"\t-Dservice.tpl.service_base_impl=" + _TPL_ROOT + "service_base_impl.ftl\n" +
-				"\t-Dservice.tpl.service_clp=" + _TPL_ROOT + "service_clp.ftl\n" +
-				"\t-Dservice.tpl.service_clp_invoker=" + _TPL_ROOT + "service_clp_invoker.ftl\n" +
-				"\t-Dservice.tpl.service_clp_message_listener=" + _TPL_ROOT + "service_clp_message_listener.ftl\n" +
-				"\t-Dservice.tpl.service_clp_serializer=" + _TPL_ROOT + "service_clp_serializer.ftl\n" +
-				"\t-Dservice.tpl.service_http=" + _TPL_ROOT + "service_http.ftl\n" +
-				"\t-Dservice.tpl.service_impl=" + _TPL_ROOT + "service_impl.ftl\n" +
-				"\t-Dservice.tpl.service_props_util=" + _TPL_ROOT + "service_props_util.ftl\n" +
-				"\t-Dservice.tpl.service_soap=" + _TPL_ROOT + "service_soap.ftl\n" +
-				"\t-Dservice.tpl.service_util=" + _TPL_ROOT + "service_util.ftl\n" +
-				"\t-Dservice.tpl.service_wrapper=" + _TPL_ROOT + "service_wrapper.ftl\n" +
-				"\t-Dservice.tpl.spring_xml=" + _TPL_ROOT + "spring_xml.ftl\n" +
-				"\t-Dservice.tpl.spring_xml_session=" + _TPL_ROOT + "spring_xml_session.ftl";
-
 			if (e instanceof ServiceBuilderException) {
 				System.err.println(e.getMessage());
 			}
 			else {
-				System.out.println(message);
+				StringBundler sb = new StringBundler(160);
+
+				sb.append("Please set these arguments. Sample values are:\n");
+				sb.append("\n");
+				sb.append("\tservice.api.dir=${basedir}/../portal-kernel/src\n");
+				sb.append("\tservice.auto.import.default.references=true\n");
+				sb.append("\tservice.auto.namespace.tables=false\n");
+				sb.append("\tservice.bean.locator.util=com.liferay.portal.kernel.bean.PortalBeanLocatorUtil\n");
+				sb.append("\tservice.build.number=1\n");
+				sb.append("\tservice.build.number.increment=true\n");
+				sb.append("\tservice.hbm.file=${basedir}/src/META-INF/portal-hbm.xml\n");
+				sb.append("\tservice.impl.dir=${basedir}/src\n");
+				sb.append("\tservice.input.file=${service.file}\n");
+				sb.append("\tservice.model.hints.configs=");
+				sb.append(StringUtil.merge(ServiceBuilderArgs.MODEL_HINTS_CONFIGS));
+				sb.append("\n");
+				sb.append("\tservice.model.hints.file=${basedir}/src/META-INF/portal-model-hints.xml\n");
+				sb.append("\tservice.osgi.module=false\n");
+				sb.append("\tservice.plugin.name=\n");
+				sb.append("\tservice.props.util=com.liferay.portal.util.PropsUtil\n");
+				sb.append("\tservice.read.only.prefixes=");
+				sb.append(StringUtil.merge(ServiceBuilderArgs.READ_ONLY_PREFIXES));
+				sb.append("\n");
+				sb.append("\tservice.resource.actions.configs=");
+				sb.append(StringUtil.merge(ServiceBuilderArgs.RESOURCE_ACTION_CONFIGS));
+				sb.append("\n");
+				sb.append("\tservice.resources.dir=${basedir}/src\n");
+				sb.append("\tservice.spring.file=${basedir}/src/META-INF/portal-spring.xml\n");
+				sb.append("\tservice.spring.namespaces=beans\n");
+				sb.append("\tservice.sql.dir=${basedir}/../sql\n");
+				sb.append("\tservice.sql.file=portal-tables.sql\n");
+				sb.append("\tservice.sql.indexes.file=indexes.sql\n");
+				sb.append("\tservice.sql.sequences.file=sequences.sql\n");
+				sb.append("\tservice.target.entity.name=${service.target.entity.name}\n");
+				sb.append("\tservice.test.dir=${basedir}/test/integration\n");
+				sb.append("\n");
+				sb.append("You can also customize the generated code by overriding the default templates with these optional system properties:\n");
+				sb.append("\n");
+				sb.append("\t-Dservice.tpl.bad_alias_names=");
+				sb.append(_TPL_ROOT);
+				sb.append("bad_alias_names.txt\n");
+				sb.append("\t-Dservice.tpl.bad_column_names=");
+				sb.append(_TPL_ROOT);
+				sb.append("bad_column_names.txt\n");
+				sb.append("\t-Dservice.tpl.bad_json_types=");
+				sb.append(_TPL_ROOT);
+				sb.append("bad_json_types.txt\n");
+				sb.append("\t-Dservice.tpl.bad_table_names=");
+				sb.append(_TPL_ROOT);
+				sb.append("bad_table_names.txt\n");
+				sb.append("\t-Dservice.tpl.base_mode_impl=");
+				sb.append(_TPL_ROOT);
+				sb.append("base_mode_impl.ftl\n");
+				sb.append("\t-Dservice.tpl.blob_model=");
+				sb.append(_TPL_ROOT);
+				sb.append("blob_model.ftl\n");
+				sb.append("\t-Dservice.tpl.copyright.txt=copyright.txt\n");
+				sb.append("\t-Dservice.tpl.ejb_pk=");
+				sb.append(_TPL_ROOT);
+				sb.append("ejb_pk.ftl\n");
+				sb.append("\t-Dservice.tpl.exception=");
+				sb.append(_TPL_ROOT);
+				sb.append("exception.ftl\n");
+				sb.append("\t-Dservice.tpl.extended_model=");
+				sb.append(_TPL_ROOT);
+				sb.append("extended_model.ftl\n");
+				sb.append("\t-Dservice.tpl.extended_model_base_impl=");
+				sb.append(_TPL_ROOT);
+				sb.append("extended_model_base_impl.ftl\n");
+				sb.append("\t-Dservice.tpl.extended_model_impl=");
+				sb.append(_TPL_ROOT);
+				sb.append("extended_model_impl.ftl\n");
+				sb.append("\t-Dservice.tpl.finder=");
+				sb.append(_TPL_ROOT);
+				sb.append("finder.ftl\n");
+				sb.append("\t-Dservice.tpl.finder_base_impl=");
+				sb.append(_TPL_ROOT);
+				sb.append("finder_base_impl.ftl\n");
+				sb.append("\t-Dservice.tpl.finder_util=");
+				sb.append(_TPL_ROOT);
+				sb.append("finder_util.ftl\n");
+				sb.append("\t-Dservice.tpl.hbm_xml=");
+				sb.append(_TPL_ROOT);
+				sb.append("hbm_xml.ftl\n");
+				sb.append("\t-Dservice.tpl.json_js=");
+				sb.append(_TPL_ROOT);
+				sb.append("json_js.ftl\n");
+				sb.append("\t-Dservice.tpl.json_js_method=");
+				sb.append(_TPL_ROOT);
+				sb.append("json_js_method.ftl\n");
+				sb.append("\t-Dservice.tpl.model=");
+				sb.append(_TPL_ROOT);
+				sb.append("model.ftl\n");
+				sb.append("\t-Dservice.tpl.model_cache=");
+				sb.append(_TPL_ROOT);
+				sb.append("model_cache.ftl\n");
+				sb.append("\t-Dservice.tpl.model_hints_xml=");
+				sb.append(_TPL_ROOT);
+				sb.append("model_hints_xml.ftl\n");
+				sb.append("\t-Dservice.tpl.model_impl=");
+				sb.append(_TPL_ROOT);
+				sb.append("model_impl.ftl\n");
+				sb.append("\t-Dservice.tpl.model_soap=");
+				sb.append(_TPL_ROOT);
+				sb.append("model_soap.ftl\n");
+				sb.append("\t-Dservice.tpl.model_wrapper=");
+				sb.append(_TPL_ROOT);
+				sb.append("model_wrapper.ftl\n");
+				sb.append("\t-Dservice.tpl.persistence=");
+				sb.append(_TPL_ROOT);
+				sb.append("persistence.ftl\n");
+				sb.append("\t-Dservice.tpl.persistence_impl=");
+				sb.append(_TPL_ROOT);
+				sb.append("persistence_impl.ftl\n");
+				sb.append("\t-Dservice.tpl.persistence_util=");
+				sb.append(_TPL_ROOT);
+				sb.append("persistence_util.ftl\n");
+				sb.append("\t-Dservice.tpl.props=");
+				sb.append(_TPL_ROOT);
+				sb.append("props.ftl\n");
+				sb.append("\t-Dservice.tpl.service=");
+				sb.append(_TPL_ROOT);
+				sb.append("service.ftl\n");
+				sb.append("\t-Dservice.tpl.service_base_impl=");
+				sb.append(_TPL_ROOT);
+				sb.append("service_base_impl.ftl\n");
+				sb.append("\t-Dservice.tpl.service_clp=");
+				sb.append(_TPL_ROOT);
+				sb.append("service_clp.ftl\n");
+				sb.append("\t-Dservice.tpl.service_clp_invoker=");
+				sb.append(_TPL_ROOT);
+				sb.append("service_clp_invoker.ftl\n");
+				sb.append("\t-Dservice.tpl.service_clp_message_listener=");
+				sb.append(_TPL_ROOT);
+				sb.append("service_clp_message_listener.ftl\n");
+				sb.append("\t-Dservice.tpl.service_clp_serializer=");
+				sb.append(_TPL_ROOT);
+				sb.append("service_clp_serializer.ftl\n");
+				sb.append("\t-Dservice.tpl.service_http=");
+				sb.append(_TPL_ROOT);
+				sb.append("service_http.ftl\n");
+				sb.append("\t-Dservice.tpl.service_impl=");
+				sb.append(_TPL_ROOT);
+				sb.append("service_impl.ftl\n");
+				sb.append("\t-Dservice.tpl.service_props_util=");
+				sb.append(_TPL_ROOT);
+				sb.append("service_props_util.ftl\n");
+				sb.append("\t-Dservice.tpl.service_soap=");
+				sb.append(_TPL_ROOT);
+				sb.append("service_soap.ftl\n");
+				sb.append("\t-Dservice.tpl.service_util=");
+				sb.append(_TPL_ROOT);
+				sb.append("service_util.ftl\n");
+				sb.append("\t-Dservice.tpl.service_wrapper=");
+				sb.append(_TPL_ROOT);
+				sb.append("service_wrapper.ftl\n");
+				sb.append("\t-Dservice.tpl.spring_xml=");
+				sb.append(_TPL_ROOT);
+				sb.append("spring_xml.ftl\n");
+				sb.append("\t-Dservice.tpl.spring_xml_session=");
+				sb.append(_TPL_ROOT);
+				sb.append("spring_xml_session.ftl");
+
+				System.out.println(sb.toString());
 			}
 
 			ArgumentsUtil.processMainException(arguments, e);
@@ -2307,12 +2396,15 @@ public class ServiceBuilder {
 		File xmlFile = new File(_hbmFileName);
 
 		if (!xmlFile.exists()) {
-			String xml =
-				"<?xml version=\"1.0\"?>\n" +
-				"<!DOCTYPE hibernate-mapping PUBLIC \"-//Hibernate/Hibernate Mapping DTD 3.0//EN\" \"http://hibernate.sourceforge.net/hibernate-mapping-3.0.dtd\">\n" +
-				"\n" +
-				"<hibernate-mapping auto-import=\"false\" default-lazy=\"false\">\n" +
-				"</hibernate-mapping>";
+			StringBundler sb = new StringBundler(5);
+
+			sb.append("<?xml version=\"1.0\"?>\n");
+			sb.append("<!DOCTYPE hibernate-mapping PUBLIC \"-//Hibernate/Hibernate Mapping DTD 3.0//EN\" \"http://hibernate.sourceforge.net/hibernate-mapping-3.0.dtd\">\n");
+			sb.append("\n");
+			sb.append("<hibernate-mapping auto-import=\"false\" default-lazy=\"false\">\n");
+			sb.append("</hibernate-mapping>");
+
+			String xml = sb.toString();
 
 			_write(xmlFile, xml);
 		}
@@ -2543,13 +2635,9 @@ public class ServiceBuilder {
 		File xmlFile = new File(_modelHintsFileName);
 
 		if (!xmlFile.exists()) {
-			String xml =
-				"<?xml version=\"1.0\"?>\n" +
-				"\n" +
-				"<model-hints>\n" +
-				"</model-hints>";
-
-			_write(xmlFile, xml);
+			_write(
+				xmlFile,
+				"<?xml version=\"1.0\"?>\n\n<model-hints>\n</model-hints>");
 		}
 
 		String oldContent = _read(xmlFile);
@@ -3277,16 +3365,21 @@ public class ServiceBuilder {
 
 		File xmlFile = new File(_springFileName);
 
-		String xml =
-			"<?xml version=\"1.0\"?>\n" +
-			"\n" +
-			"<beans\n" +
-			"\tdefault-destroy-method=\"destroy\"\n" +
-			"\tdefault-init-method=\"afterPropertiesSet\"\n" +
-			_getSpringNamespacesDeclarations() +
-			"\txmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\"\n" +
-			"\txsi:schemaLocation=\"" + _getSpringSchemaLocations() + "\">\n" +
-			"</beans>";
+		StringBundler sb = new StringBundler(11);
+
+		sb.append("<?xml version=\"1.0\"?>\n\n");
+		sb.append("<beans\n");
+		sb.append("\tdefault-destroy-method=\"destroy\"\n");
+		sb.append("\tdefault-init-method=\"afterPropertiesSet\"\n");
+		sb.append(_getSpringNamespacesDeclarations());
+		sb.append("\txmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\"");
+		sb.append("\n");
+		sb.append("\txsi:schemaLocation=\"");
+		sb.append(_getSpringSchemaLocations());
+		sb.append("\">\n");
+		sb.append("</beans>");
+
+		String xml = sb.toString();
 
 		if (!xmlFile.exists()) {
 			_write(xmlFile, xml);