+++ portal-impl/src/com/liferay/portal/spring/hibernate/LastSessionRecorderHibernateTransactionManager.java
@@ -14,17 +14,8 @@
 
 package com.liferay.portal.spring.hibernate;
 
-import com.liferay.portal.kernel.util.CentralizedThreadLocal;
-import com.liferay.portal.kernel.util.InitialThreadLocal;
-import com.liferay.portal.kernel.util.ReflectionUtil;
-
-import java.lang.reflect.Field;
-import java.lang.reflect.Modifier;
-
-import org.springframework.core.NamedThreadLocal;
 import org.springframework.orm.hibernate3.HibernateTransactionManager;
 import org.springframework.orm.hibernate3.SessionHolder;
-import org.springframework.transaction.support.TransactionSynchronizationManager;
 
 /**
  * @author Shuyang Zhou
@@ -35,8 +26,7 @@ public class LastSessionRecorderHibernateTransactionManager
 	@Override
 	protected Object doGetTransaction() {
 		SessionHolder sessionHolder =
-			(SessionHolder)TransactionSynchronizationManager.getResource(
-				getSessionFactory());
+			SpringHibernateThreadLocalUtil.getResource(getSessionFactory());
 
 		if (sessionHolder != null) {
 			LastSessionRecorderUtil.setLastSession(sessionHolder.getSession());
@@ -47,38 +37,10 @@ public class LastSessionRecorderHibernateTransactionManager
 
 	static {
 		try {
-			Field nameField = ReflectionUtil.getDeclaredField(
-				NamedThreadLocal.class, "name");
-
-			for (Field field : ReflectionUtil.getDeclaredFields(
-					TransactionSynchronizationManager.class)) {
-
-				if (Modifier.isStatic(field.getModifiers()) &&
-					ThreadLocal.class.isAssignableFrom(field.getType())) {
-
-					ThreadLocal<Object> threadLocal =
-						(ThreadLocal<Object>)field.get(null);
-
-					Object value = threadLocal.get();
-
-					if (threadLocal instanceof NamedThreadLocal) {
-						threadLocal = new InitialThreadLocal<>(
-							(String)nameField.get(threadLocal), null);
-					}
-					else {
-						threadLocal = new CentralizedThreadLocal<>(false);
-					}
-
-					if (value != null) {
-						threadLocal.set(value);
-					}
-
-					field.set(null, threadLocal);
-				}
-			}
+			Class.forName(SpringHibernateThreadLocalUtil.class.getName());
 		}
-		catch (Exception e) {
-			throw new ExceptionInInitializerError(e);
+		catch (ClassNotFoundException cnfe) {
+			throw new ExceptionInInitializerError(cnfe);
 		}
 	}