+++ modules/sdk/gradle-plugins-js-module-config-generator/CHANGELOG.markdown
@@ -1,6 +1,6 @@
 # Liferay Gradle Plugins JS Module Config Generator Change Log
 
-## 1.0.31 - 2016-08-26
+## 1.0.31 - 2016-08-27
 
 ### Changed
 - [LPS-67023]: Update the [Liferay Gradle Plugins Node] dependency to version
@@ -19,8 +19,15 @@
 - [LPS-68485]: The up-to-date check for `ConfigJSModulesTask` tasks is incorrect
 (files are modified in-place) and it has been disabled.
 
+## 2.0.2 - 2016-10-06
+
+### Changed
+- [LPS-68564]: Update the [Liferay Gradle Plugins Node] dependency to version
+1.2.0.
+
 [Liferay Gradle Plugins Node]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins-node
 [LPS-66906]: https://issues.liferay.com/browse/LPS-66906
 [LPS-67023]: https://issues.liferay.com/browse/LPS-67023
 [LPS-67573]: https://issues.liferay.com/browse/LPS-67573
 [LPS-68485]: https://issues.liferay.com/browse/LPS-68485
+[LPS-68564]: https://issues.liferay.com/browse/LPS-68564