+++ modules/sdk/gradle-plugins/CHANGELOG.markdown
@@ -9,7 +9,7 @@
 - [LPS-67658]: Compile the plugin against Gradle 2.14 to make it compatible with
 both Gradle 2.14+ and Gradle 3.0.
 
-## 2.0.11 - 2016-08-22
+## 2.0.11 - 2016-08-23
 
 ### Changed
 - [LPS-67352]: Update the [Liferay Source Formatter] dependency to version
@@ -289,6 +289,37 @@ to version 1.0.13.
 - [LPS-68334]: Update the [Liferay Portal Tools Service Builder] dependency to
 version 1.0.128.
 
+## 3.0.0 - 2016-10-06
+
+### Changed
+- [LPS-67352]: Update the [Liferay Source Formatter] dependency to version
+1.0.291.
+- [LPS-67573]: Make most methods private in order to reduce API surface.
+
+### Removed
+- [LPS-66396]: The task classes `BuildThumbnailsTask` and `CompileThemeTask` in
+`com.liferay.gradle.plugins.tasks` are no longer available. Please use
+[Liferay Gradle Plugins Theme Builder] instead.
+- [LPS-67573]: To reduce the number of plugins applied to a project and improve
+performance, plugins in `com.liferay.gradle.plugins.internal` are no longer
+applied via `apply plugin`.
+
+## 3.0.1 - 2016-10-06
+
+### Changed
+- [LPS-68415]: Update the [Liferay Source Formatter] dependency to version
+1.0.292.
+
+## 3.0.2 - 2016-10-06
+
+### Changed
+- [LPS-68564]: Update the [Liferay Gradle Plugins Gulp] dependency to version
+2.0.1.
+- [LPS-68564]: Update the [Liferay Gradle Plugins JS Module Config Generator]
+dependency to version 2.0.2.
+- [LPS-68564]: Update the [Liferay Gradle Plugins JS Transpiler] dependency to
+version 2.0.1.
+
 [Gradle Bundle Plugin]: https://github.com/TomDmitriev/gradle-bundle-plugin
 [Liferay Ant BND]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/ant-bnd
 [Liferay CSS Builder]: https://github.com/liferay/liferay-portal/tree/master/modules/util/css-builder
@@ -297,6 +328,7 @@ version 1.0.128.
 [Liferay Gradle Plugins JS Transpiler]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins-js-transpiler
 [Liferay Gradle Plugins Service Builder]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins-service-builder
 [Liferay Gradle Plugins Source Formatter]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins-source-formatter
+[Liferay Gradle Plugins Theme Builder]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins-theme-builder
 [Liferay Javadoc Formatter]: https://github.com/liferay/liferay-portal/tree/master/modules/util/javadoc-formatter
 [Liferay Lang Builder]: https://github.com/liferay/liferay-portal/tree/master/modules/util/lang-builder
 [Liferay Portal Tools Service Builder]: https://github.com/liferay/liferay-portal/tree/master/modules/util/portal-tools-service-builder
@@ -307,10 +339,12 @@ version 1.0.128.
 [Liferay XML Formatter]: https://github.com/liferay/liferay-portal/tree/master/modules/util/xml-formatter
 [LPS-58672]: https://issues.liferay.com/browse/LPS-58672
 [LPS-61099]: https://issues.liferay.com/browse/LPS-61099
+[LPS-66396]: https://issues.liferay.com/browse/LPS-66396
 [LPS-66853]: https://issues.liferay.com/browse/LPS-66853
 [LPS-66906]: https://issues.liferay.com/browse/LPS-66906
 [LPS-67023]: https://issues.liferay.com/browse/LPS-67023
 [LPS-67352]: https://issues.liferay.com/browse/LPS-67352
+[LPS-67573]: https://issues.liferay.com/browse/LPS-67573
 [LPS-67658]: https://issues.liferay.com/browse/LPS-67658
 [LPS-67766]: https://issues.liferay.com/browse/LPS-67766
 [LPS-67804]: https://issues.liferay.com/browse/LPS-67804
@@ -322,5 +356,7 @@ version 1.0.128.
 [LPS-68165]: https://issues.liferay.com/browse/LPS-68165
 [LPS-68297]: https://issues.liferay.com/browse/LPS-68297
 [LPS-68334]: https://issues.liferay.com/browse/LPS-68334
+[LPS-68415]: https://issues.liferay.com/browse/LPS-68415
 [LPS-68485]: https://issues.liferay.com/browse/LPS-68485
 [LPS-68504]: https://issues.liferay.com/browse/LPS-68504
+[LPS-68564]: https://issues.liferay.com/browse/LPS-68564